Implement basic XMPP client.
This adds a very basic XMPP client based on gloox, which just establishes
a connection and can disconnect again.  Handling of messages is not
implemented, as that needs to be done in subclasses as appropriate.

We also have unit tests that verify it actually works by connecting
to the live server on chat.xaya.io.