Add threading to XmppClient.
Make the XmppClient itself run a separate thread for receiving messages,
and implement proper synchronisation for other threads that also want
to access the underlying gloox::Client, e.g. for sending messages.

This is a bit ugly with locks and sleeps, but works for now.  We will need
this or a similar implementation later anyway, because both the Charon
client and server pieces will need to listen to XMPP as well as send
messages triggered from the "outside" on other threads.