Update protocol for handshake with presence.
Adapt the description of the protocol, so that the initial handshake
uses also directed presence.  By exchanging a pair of directed presence
stanzas, client and server will subscribe (temporarily) to each other's
presence.  This has the advantage that they will get notified by the
XMPP server if one of them becomes unavailable, and can then react to that
situation accordingly (e.g. the client can select another server resource
to replace the one that became unavailable).