Enforce TLS on the connection.
Require TLS on the connection made by gloox.  It was disabled before
due to an issue with TLS 1.3 in GnuTLS, but enforcing TLS works fine
if we "downgrade" to TLS 1.2 server-side.  Thus we should enable
TLS on connections.

Note that we are not yet actually enforcing validity of server certs,
so a connection could in theory be MitM attacked.  We need to do that
at some later stage as well.