Update notifications for the Charon client.
Add support to notifications to the client.  When enabling certain
notification types, the client will make sure to select a server that
advertises them.  It will then also subscribe to the server's nodes,
and keep track of the current state for each type by listening to the
server broadcasts.

With this, it implements a waitforchange-like interface that can be used
by the actual RPC server in the end (but works purely locally).