Add Dockerfile for Charon (+libxayagame).
The new Dockerfile allows building a Debian-based image with Charon
built and installed, as well as libxayagame.  The latter is not
strictly a dependency of Charon, but most users will want to combine
Charon and libxayagame for e.g. GSP images.  Thus the Charon images built
this way will essentially be "Xaya platform SDK development" images.