Support --methods_exclude for utilities.
Add a new argument, --methods_exclude, for the utility binaries.  This
can be used to exclude a comma-separated list of methods from being
exposed.

Also extends the integration tests by one that is meant to test the
method selection arguments specifically.

For now, this is not very useful, as one can just leave out those methods
from --methods instead.  But we will use this framework now to also support
specifying the list of methods by passing the jsonrpccpp specification file,
where it may be useful to exclude methods.