Turn on TLS CA verification.
Make a connection to the XMPP server fail if gloox signals that the
server's TLS certificate is not trusted.  It is possible to set an
explicit root certificate that is expected to have signed the
server cert in the API.  Otherwise the system's default trust store
will be used.

For the local test environment, we use a self-signed certificate
that is also checked in, used in the Docker image that runs the
test server, and has a corresponding CA root certificate file that
the tests use.