Report version string by server.
For instantiating a server, we now need to provide a version string
in addition to the RPC backend.  The version string should indicate the
particular "version" (e.g. consensus forks, interface) of the backend
provided.

This can later be used by clients to ensure the server they are connected
to provides the interface and rules they want.