Add sleeps in tests.
When checking that the server cleans up PubSub nodes, wait a bit before
checking to make sure the deletion has gone through.  The destructor itself
just sends the request but does not wait for confirmations (as they are
not necessary unless we are testing it).

Similarly make sure that the client had time to receive the disconnect
message from the server when we check server reselection.