Allow choosing server for integration tests.
Update the integration test code to allow choosing the XMPP server
and environment to run the tests against.  It is by default the new
localhost environment, but can also be set to the production
chat.xaya.io server that was used before.  This is especially useful
to verify the server configuration itself for use with Charon.