Fix deadlock with multiple servers.
When multiple servers with notifications were running, the client could
deadlock when processing the initial subscriptions.  This was because
it kept hold on its mut lock while waiting for the subscriptions to be
done, but processing the second server's presence stanza would also
wait for the lock; hence the subscription confirmations would never
be processed anymore.