Decouple pubsub from server notifications.
Decouple the ServerNotification instances held by a Server object
from the attached PubSub.  This allows the notifications to be added
and the waiter threads to be running even before the XMPP client is
connected, or while it is disconnected and later reconnected.

Only the pubsub instance and associated nodes are now tied to the
XMPP connection itself.

With this, it is possible to use AddNotification before the Server
instance is Connect'ed.