Notify XmppClient subclasses about disconnects.
This adds a new method XmppClient::HandleDisconnect, which subclasses
can override to get notified about disconnects (both when done explicitly
and when forced by the server).  We use this in Server to clean up
notifications for now (but will later use that in Client as well).

Also streamlines how the XmppClient cleans up a potential pubsub instance
when disconnecting.