Make charon-server reconnect automatically.
Let the charon-server main loop reconnect the server (or attempt to
at least) periodically when it gets disconnected.  This ensures that
the server will recover eventually from a temporary XMPP server failure
or restart.

This also adds a new integration test xmpp_restart.py, which has
to be run manually as the testing XMPP server needs to be restarted
in the middle of the test.  It tests that a started Charon server
behaves fine even if the XMPP server is restarted in the middle
of doing things, and that it will reconnect by itself and be able
to serve clients again.