Use shared pointers for UpdatableState.
The UpdatableState used in tests was delicate to use, because it had
to be ensured that the underlying server and waiter threads are destroyed
before the state instances themselves.

Instead of requiring this, we now just use std::shared_ptr instances
to hold the state and from external code.  This allows us to make sure
all waiter threads keep a shared pointer as well, and thus ensure that
the state stays around long enough in every case.