Back off when waiter call fails.
When the WaiterThread's call fails, do not attempt to do another
right away, but instead sleep for a certain backoff time.  This ensures
that we will not spam millions of failing RPCs when e.g. the GSP is down,
but still try from time to time and recover the connection when it gets
back up.