Allow running the test env as a cluster.
Extend the test environment scripts so that the test server can optionally
be run as a cluster of two nodes rather than a single node.  This may
help to spot more interoperability issues with ejabberd in clustered mode
as well as the actual cluster configuration necessary for Charon to run.