Wait for node deletion in test.
In the unit test, introduce a small sleep between when a server is
disconnected (and its pubsub nodes are deleted) and when the server is
connected again (and new nodes are created).

If we do not do this, then ejabberd in a clustered setting may return
errors when operating on the nodes due to the cluster-wide database
locking it does for node operations.