Keep around XmppClient instance for Client.
In the Charon Client, keep around the internal XmppClient / Impl
instance for the whole lifetime of the instance.  This allows
to connect / disconnect it as desired, and it will keep on working.

The enabled notifications are now just stored in the Impl instance
together with the states, there is no need for a separate list of
notifications anymore.