Fix client priority to -1.
The Charon client will always want to connect with negative priority,
since it should never receive any messages to the bare JID.  It only
ever asks for information and then receives the response to its
full JID.

Thus there is no need to let a user choose the priority when connecting
the Client instance.  Instead, it just always connects with priority -1.