Refactor TryEnsureFullServerJid handling.
This changes the way that TryEnsureFullServerJid works in the Charon
client.  Instead of just silently doing its thing, it now returns the
full server JID or nothing if the connection failed.  This allows us
to simplify the actual "user" methods (wait for change and RPC call
forwarding), and also moves the locking down into that function itself
where we have better control.