Enforce maximum payload size for XML data.
Add a constant (set to 64 MiB) for the maximum allowed length of an
XML payload.  This is used to protect against out-of-memory DoS
attacks, especially when we add compression later.

64 MiB is small enough to not be an issue for modern systems to have
in memory (even two or three copies), while it should be more than
enough for the payloads we expect to process through Charon or XMPP
in general.