Use shared library for building utils.
For building the charon utils (charon-client, charon-server), use
a noinst shared library to hold common code (for now, just the
method selection).

We will extend the library to support running most of the "main"
code for the utils directly, so that it can be used as well from
external code.