Move main logic for charon-client to utils lib.
This moves out most of the logic for the charon-client binary from the
main file to the new utils library, and just calls the library from
the main file.

With this, we can make better use of this code in the future from
alternative entry points, e.g. to directly embed the logic into
other applications.