Include CA root in Docker images.
Update the Dockerfiles for Charon, to build gloox with the necessary
patches to make TLS certificates with GnuTLS work properly.  We also
include the letsencrypt root certificate in /usr/local/share/letsencrypt.pem
for convenience (this is the root CA that the official XMPP server at
chat.xaya.io uses).