Install CA files with Charon.
Instead of just distributing the Charon test server and the chat.xaya.io
letsencrypt certificates with the Charon sources, install them into the
pkgdatadir (typically /usr/local/share/charon).

These files are useful for other applications that need Charon, e.g.
for running their own tests or distributing as the --cafile to use
when talking to the production server.